New updates for testing.
git-svn-id: file:///home/git/hedgehog.fhcrc.org/bioconductor/trunk/madman/Rpacks/affxparser@17156 bc3139a8-67e5-0310-9ffc-ced21a209358

+++ src/fusion_sdk/calvin_files/utils/src/AffymetrixGuid.cpp
@@ -44,14 +44,14 @@ using namespace affymetrix_calvin_utilities;
 void InitializeRandomNumberGenerator()
 {
 	// Get the current time as a starting point for a seed.
-	unsigned int seed = (unsigned int) time(NULL);
+	u_int64_t seed = (u_int64_t) time(NULL);
 
 	// Multiply the time by its address and the process id.
-	seed *= (unsigned int)(&seed);
-	seed *= getpid();
+	seed *= (u_int64_t)(&seed);
+	seed *= (u_int64_t)getpid();
 
 	// Initialize the random number generator.
-	srand(seed);
+	srand((unsigned int)seed);
 }
 
 /*
